Moving os_ident to impacket.examples
