+++ impacket/examples/os_ident.py
